fix: incoming msg processing halts
In async process_messages we have a while-true loop that dequeues
incoming messages from mqtt. If we have one of the error cases,
such as malformed messages, a topic-validation mismatch, or most
likely a message type that isn't an object update or user update,
currently most often a chat message, this will return out of the
while loop and no further messages will be processed. Fixed by
changing return to continue in loop